Don't generate compat.h - use preprocessor directives instead
This simplifies the build system and avoids the need to run "make clean"
before compiling for another kernel.

Generated compat.h is only useful for changes that are not reflected in
the kernel version.  This can happen in the beginning of a development
cycle for the kernel.  However, we cannot commit to supporting obsolete
development snapshots.

As for the current development cycle, there is always a possibility to
detect it through the use of preprocessor directives, if the changes are
significant enough to require code adjustments in ndiswrapper.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2583 388e3070-e2af-4704-926e-a220f121eef5