Consolidate all sanity checks, add check for wireless extensions
As the build system is not noisy on Linux 2.6, it's better to conduct
all sanity checks before compilation.  Improve messages for visibility,
readability and correctlness.  Add a warning about wireless extension
support.

Remove a warning in ntoskernel.h - it's triggered for every source file.
And it's not even correct, as it doesn't check for CONFIG_NET_RADIO,
which is what we should be checking on Linux 2.6.16.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2635 388e3070-e2af-4704-926e-a220f121eef5