"make install" should not call Linux build unless the module is missing
Otherwise, not having some options such as DEBUG on the "make install"
command line that were on the "make all" command line would trigger a
rebuild as root, which is not what the user means. It might also create
root owned files and directories, which would interfere with future
builds or with removal of the sources.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2839 388e3070-e2af-4704-926e-a220f121eef5