Fix error handling in ndis_start_device()
ndis_remove_device() is an overkill.  It frees the network device and
causes panic down the road.  We don't allocate the network device in
ndis_start_device(), so we shouldn't be freeing it.

Use mp_halt() instead, which is a counterpart to mp_init().  If an error
happens after register_netdev(), call unregister_netdev() and set
wnd->max_tx_packets to 0 so that unregister_netdev() won't be called
again.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2877 388e3070-e2af-4704-926e-a220f121eef5