Don't use usb_init_urb(), it resets refcount to 1
usb_init_urb() should not be called on URBs allocated by
usb_alloc_urb().  It sets the URB reference count to 1 regardless of
what it was before.

When unloading ndiswrapper, the refcount may become 2.  Setting it to 1
lead to freeing the URB twice.  When freeing the USB for the second
time, the reference counter is decremented in freed memory, causing
random kernel crashes.


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@2948 388e3070-e2af-4704-926e-a220f121eef5