Fix backward logic in iw_set_essid()
Instead of coding for the old, buggy WE-19 API and adding a workaround
for WE-20 and later, code for the current API and work around bugs in
the older ones.

Scanning ESSID forward until the first 0 can lead to accessing the
memory outside the ESSID buffer.  Only check for one final 0 - that's
the bug we are working around.

Don't copy ESSID unnecessarily (already fixed in ndis6wrapper).


git-svn-id: svn://svn.code.sf.net/p/ndiswrapper/code@3079 388e3070-e2af-4704-926e-a220f121eef5